modified function from React to JS
