+++ src/App.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 // import axios from 'axios';
 import './App.css';
-import EventsFinder from './components/EventsFinder';
+import EventsTest from './components/EventsTest';
 
 function App() {
   return (
     <div className="App">
-      <EventsFinder />
+      <EventsTest />
     </div>
   );
 }
